Finished CRUD on Frontend and Refactored Backend

+++ e-commerce-backend/db/migrations/20171212095047_create_animes_table.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
     table.string('photo_url');
     table.string('title');
     table.string('rating');
-    table.integer('price');
+    table.decimal('price');
     table.boolean('on_sale').defaultsTo(false);
     table.boolean('in_cart').defaultsTo(false);
   });
